libadmin: parse rxstat_* command line args with libcmd
Use the libcmd library to parse rxstat_* command line arguments, instead
of custom code in each program. The positional argument syntax is
unchanged by this commit. Optional command line switches for the
arguments are now supported, as a feature of libcmd.

This also fixes the incorrect usage messages in several of the commands.

Change-Id: I5df9b9a439bc12e7c1090abc9713e5245d666c1b
Reviewed-on: https://gerrit.openafs.org/12374
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>