Update debian/rules for new upstream version
Some packages are gone, with klog and the pam module and such.

We now have a more complicated krb5 setup, as we use MIT krb5 for
libkrb5 functionality but also pull in libroken and libhcrypto
from heimdal.

Now that upstream supports passing CFLAGS/etc. to configure, we
can pass them directly instead of smushing them into CC the way
we started doing when hardening-wrapper went away.