Work on fixing test failures on !amd64
One is reproducible on i386, and the result of an extra sizeof()
in a loop boundary condition, so it stops at sizeof(size_t) rather
than the size of the array (8).  The other only appears on the
buildds for some more exotic architectures, so run that test
verbosely and see if we can get enough information to track
things down.