bozo: Do not create client directory and symlinks
Each time the bosserver starts, it checks for the presence of the client
configuration directory and the CellServDB and ThisCell files within it.
When not found, the bosserver creates the client cell configuration
directory. When the CellServDB and ThisCell files are not present in the
client configuration directory, the bosserver creates symlinks to the
server's CellServDB and ThisCell files.  This feature of the bosserver
was a convenience when older versions of vos and pts required a client
configuration directory.

However, modern deployments tend to use packaging, with separate client
and server packages.  The client configuration directory and files in it
are typically managed by the client packaging.  Currently, packagers
must work around these symlinks created by the bosserver.  Additionally,
the CellServDB and ThisCell symlinks are hazardous since writing to the
client-side configuration will overwrite the server configuration.

This commit removes the creation the client configuration directory and
the CellServDB and ThisCell symlinks during bosserver startup.  This
change is intended to decouple the server from the client, help to avoid
overwriting the server configuration, and avoid requiring client artifacts
on a server.

Change-Id: I8aba81185a52e6792d36ae28a430ebe1c23919d2
Reviewed-on: https://gerrit.openafs.org/12586
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>